Refactor Calendar to a separate folder
