+++ package.json
@@ -9,8 +9,10 @@
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.3.2",
         "@testing-library/user-event": "^7.1.2",
+        "@types/lodash": "^4.14.160",
         "@types/mongodb": "^3.5.26",
         "@types/mongoose": "^5.7.36",
+        "@types/react-bootstrap": "^0.32.22",
         "@types/react-router-dom": "^5.1.5",
         "@types/socket.io": "^2.1.11",
         "@types/socket.io-client": "^1.4.33",
@@ -37,6 +39,7 @@
         "jest-resolve": "24.9.0",
         "jest-watch-typeahead": "0.4.2",
         "less-loader": "^6.2.0",
+        "lodash": "^4.17.20",
         "mini-css-extract-plugin": "0.9.0",
         "mongodb": "^3.6.0",
         "mongoose": "^5.10.0",
@@ -49,6 +52,7 @@
         "postcss-safe-parser": "4.0.1",
         "react": "^16.13.1",
         "react-app-polyfill": "^1.0.6",
+        "react-bootstrap": "^1.3.0",
         "react-dev-utils": "^10.2.1",
         "react-dom": "^16.13.1",
         "react-router-dom": "^5.2.0",
